Make user_token and user_email configurable
Rather than forcing the API to return a response with user_token
and user_email in order for the session to be persisted, you should
be able to configure what properties to look for in the response. This
adds tokenAttributeName and emailAttributeName that may be configured
to reflect the API that Simple Auth is interacting with.