improved authorizer api
Authorizers will now receive a block as the single argument that they
call with any authorization data that's specific to the authorization
mechanism they implement. That way the authorizers themselves do not
depend on e.g. jQuery anymore (which they previously did as they used
jqXHR.setRequestHeader).

The session service also gets a new authorize method that takes the name
of the authorizer to use as registered in the container and the block to
call when the session is authenticated and authorization data is present.