Fix issue where store is emptied on page refresh
Related #990 #1090.

Early in the initialization of this object, `setUnknownProperty` gets called with `__OWNER__` and `__NAME_KEY__` (system properties?). This happens before `this.content` is present. The side effect is that `null` is persisted, which trashes the data within our store. This fix checks to make sure `this.content` exists before persisting. 

Related discussion #990, #1090