session-stores/cookie: Convert private volatile properties to methods
Using native getters would work too, but only in some newer versions of Ember. Older versions will have `this` assigned to the wrong thing and are not able to access `this.get` due to that. Using a regular method should work fine though and since these properties are private we can change the signature without it having to be a breaking change.